requirements.txt
aiohttp==3.9.5
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
brotlipy==0.7.0
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5d968ni_yn/croot/certifi_1671487774636/work/certifi
cffi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_ab19r4bji3/croot/cffi_1670423206034/work
cfgv==3.3.1
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
conda==23.1.0
conda-content-trust @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_11146a2b-93c2-444c-a378-ad4fac363e991s0r1hnp/croots/recipe/conda-content-trust_1658126383571/work
conda-package-handling @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5f3f459f9v/croot/conda-package-handling_1672865025324/work
conda_package_streaming @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_58gpsa_6af/croot/conda-package-streaming_1670508144037/work
cryptography @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_8c_qoky_1b/croot/cryptography_1673298759921/work
distlib==0.3.6
Django==4.2.3
django-rest-framework==0.1.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
Faker==22.5.1
filelock==3.11.0
frozenlist==1.4.1
identify==2.5.22
idna @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_771olrhiqw/croot/idna_1666125579282/work
mpmath==1.3.0
multidict==6.0.5
nodeenv==1.7.0
numpy==1.26.4
openpyxl==3.1.2
packaging==23.1
platformdirs==3.2.0
playsound==1.3.0
pluggy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_croot-w6jyveby/pluggy_1648109277227/work
pre-commit==3.2.2
psycopg2-binary==2.9.6
pycosat @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_e9xhel3kyj/croot/pycosat_1666805516297/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
pynput==1.7.6
pyobjc-core==10.1
pyobjc-framework-ApplicationServices==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-CoreText==10.1
pyobjc-framework-Quartz==10.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
PySocks @ file:///Users/ktietz/ci_310/pysocks_1643961536721/work
python-dateutil==2.8.2
PyYAML==6.0
replicate==0.11.0
requests @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_b15e28ca-768e-4951-b4c6-8a7476e43183hrjpnkg9/croots/recipe/requests_1657734638595/work
rest-framework-simplejwt==0.0.2
ruamel.yaml @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_aeh5mqcw49/croot/ruamel.yaml_1666304555976/work
ruamel.yaml.clib @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_f64xdg2rww/croot/ruamel.yaml.clib_1666302244208/work
ruff==0.1.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
snakeviz==2.2.0
sqlparse==0.4.4
stripe==7.12.0
sympy==1.12
tenacity==8.3.0
toolz @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_362wyqvvgy/croot/toolz_1667464079070/work
tornado==6.3.2
tqdm @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_637jfu39ar/croots/recipe/tqdm_1664392693982/work
tuna==0.5.11
typing_extensions==4.7.1
urllib3 @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_8fvc4z_gis/croot/urllib3_1673575619550/work
uv==0.1.41
virtualenv==20.21.0
vobject==0.9.7
voyageai==0.2.2
yarl==1.9.4
zstandard @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_99wz2rcox4/croots/recipe/zstandard_1663827384884/work
